+++ src/maths/vec3f/operations.c
@@ -12,7 +12,7 @@
 
 #include "vec3f.h"
 
-void	vec3f_mul_unit2(t_vec3f *a, float f)
+void	vec3f_mul_unit2(t_vec3f *a, double f)
 {
 	vec3f_set(a, a->x * f, a->y * f, a->z * f);
 }
@@ -27,7 +27,7 @@ void	vec3f_add2(t_vec3f *a, t_vec3f *b)
 	vec3f_set(a, a->x + b->x, a->y + b->y, a->z + b->z);
 }
 
-void	vec3f_add_unit2(t_vec3f *a, float b)
+void	vec3f_add_unit2(t_vec3f *a, double b)
 {
 	vec3f_set(a, a->x + b, a->y + b, a->z + b);
 }
